==> code/4-2_dataset_10-Q_sections.ipynb <==
############# This code creates *crsp_comp_edgar_section.csv* for sectional analysis in 10-Q ###################

############### Import packages
import os, numpy as np, pandas as pd, time, glob, re, math, statsmodels.api as sm, patsy as ps
from tqdm import tqdm
from time import process_time
from datetime import datetime
from datetime import date
from openpyxl import load_workbook
from patsy import dmatrices

############### Set working directory to parent directory
if os.getcwd() != 'F:\\github\\narrative_conservatism\\code':
    os.chdir('F:\\github\\narrative_conservatism\\code')
    
############### Set pandas column printing constraint
pd.set_option('display.max_columns', None)

########################################################################################
############ Concatenate and prepare merge: ID_DATA and TEXT_DATA ######################
########################################################################################

############## Define a function to concatenate all csv files with file name that matches a certain pattern into one data frame
def concatenate (indir, file_name_match):
    os.chdir(indir)
    file_list = glob.glob(file_name_match)
    df_list = list()
    colnames = pd.read_csv(file_list[0], header = None).loc[0]
    
    for filename in file_list:
        # print(filename)
        df = pd.read_csv(filename, low_memory = False)
        df_list.append(df)

    df_concat = pd.concat(df_list, axis = 0)
    # df_concat.columns = colnames
    return df_concat

############## Concatenate id_data and text_data files and create two data frames
text_data = concatenate('..\\filings', 'text_data_section_' + '*.csv')

############## Calculate tone : tone = (n_pos - n_negation - n_neg)/nw
text_data['tone_mda'] = (text_data['n_pos_mda'] - text_data['n_negation_mda'] - text_data['n_neg_mda'])/text_data['nw_mda']
text_data['tone_note'] = (text_data['n_pos_note'] - text_data['n_negation_note'] - text_data['n_neg_note'])/text_data['nw_note']

############## Save text_data dataframe into local file text_data_10-Q.csv
text_data.to_csv('..\\filings\\text_data_section.csv', index = 0)

print('Number of 10-Q sections parsed: ' + str(text_data.shape[0]))

########################################################################################
############### Merge SECTION_DATA with CRSP_COMPUSTAT_EDGAR_10-Q ######################
########################################################################################

crsp_comp_edgar = pd.read_csv(r'..\filings\crsp_comp_edgar_ibes_seg_10-Q.csv')

############## prepare merge: ID_DATA
crsp_comp_edgar_section = pd.merge(crsp_comp_edgar, text_data, on = ['accnum'], how = 'inner', validate = '1:1')
print('number of observations after merging with section data: ' + str(crsp_comp_edgar_section.shape[0]))
crsp_comp_edgar_section

########################################################################################
############################### Variable Creation ######################################
########################################################################################

############################## Main Variables ##########################################
######## NW: natural log of 1 + total number of words in the document
crsp_comp_edgar_section['NW_MDA'] = np.log(1 + crsp_comp_edgar_section['nw_mda'])
crsp_comp_edgar_section['NW_NOTE'] = np.log(1 + crsp_comp_edgar_section['nw_note'])

######## TONE: number of net positive words (n_pos - n_neg - n_negations) per 1000 total words
crsp_comp_edgar_section['TONE_MDA'] = crsp_comp_edgar_section['tone_mda']*1000
crsp_comp_edgar_section['TONE_NOTE'] = crsp_comp_edgar_section['tone_note']*1000

########################################################################################
############################### Variable Screening #####################################
########################################################################################

## Change SIC to str
crsp_comp_edgar_section['SIC'] = crsp_comp_edgar_section['SIC'].astype(str)

########## Drop MDAs (firm-quarter) that contain number of words less than 1% threshold
# nwq01 = crsp_comp_edgar_section['nw_mda'].quantile(.01)
# print('number of words in MDA, 1% quantile: ' + str(nwq01))
del_word01 = crsp_comp_edgar_section.loc[crsp_comp_edgar_section['nw_mda'] < 100].shape[0]
print('number of MDAs that contain total words less than 1% threshold: ' + str(del_word01))
crsp_comp_edgar_section = crsp_comp_edgar_section.loc[crsp_comp_edgar_section['nw_mda'] >= 100]

########## Drop NOTES (firm-quarter) that contain number of words less than 1% threshold
# nwq01 = crsp_comp_edgar_section['nw_note'].quantile(.01)
# print('number of words in NOTES, 1% quantile: ' + str(nwq01))
del_word01 = crsp_comp_edgar_section.loc[crsp_comp_edgar_section['nw_note'] < 100].shape[0]
print('number of NOTES that contain total words less than 1% threshold: ' + str(del_word01))
crsp_comp_edgar_section = crsp_comp_edgar_section.loc[crsp_comp_edgar_section['nw_note'] >= 100]

############## Inspect sample size after variable screening
print('Number of firm-quarters after variable screening: ' + str(crsp_comp_edgar_section.shape[0]))

############## Save merged crsp_comp_edgar_section to csv file
crsp_comp_edgar_section.to_csv('..\\filings\\crsp_comp_edgar_section.csv', index = 0)

########################################################################################
############### Table 1: Summary Statistics and Correlation Matrix #####################
########################################################################################

############# Table 1 Panel A: Summary statistics for selected variables
######### Variable groups:
# 1st line: textual variables, generally consistent with LM's summary statistics
# 2nd line: fundamental variables (main)
# 3rd line: abtone
selected_vars = crsp_comp_edgar_section[['NW','nw', 'NW_MDA','nw_mda', 'NW_NOTE','nw_note', 'TONE', 'TONE_MDA', 'TONE_NOTE', 'TLAG', \
                                   'RET', 'NEG', 'SIZE', 'MTB', 'LEV' \
                                 # 'AGE', 'age', 'EARN', 'STD_RET', 'STD_EARN', 'LOSS', 'DEARN' \
                                 ]]

T1PA = selected_vars.describe().transpose() 

############# Summary statistics for all raw and processed variables
full_summary = crsp_comp_edgar_section.describe().transpose()

T1PA
